removed trackHelper mixins and created a pure function for each of them
